feat: added switching network and wrong network banner
